@@ -1,51 +1,51 @@
-import {Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild} from '@angular/core';
-import {Coord} from '../../model/weather/coord';
-import {HourlyWeatherResponse} from '../../model/weather/hourly-weather-response';
-import {WeatherService} from '../weather.service';
-import {ChartDataSets, ChartOptions, ChartType} from 'chart.js';
-import {BaseChartDirective, Color, Label} from 'ng2-charts';
-import {finalize} from 'rxjs/operators';
-import {DatePipe} from '@angular/common';
-import {WeatherResponse} from '../../model/weather/weather-response';
-import {ResponseStatus} from '../../model/response-status.enum';
+import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
+import { Coord } from '../../model/weather/coord';
+import { HourlyWeatherResponse } from '../../model/weather/hourly-weather-response';
+import { WeatherService } from '../weather.service';
+import { ChartDataSets, ChartOptions, ChartType } from 'chart.js';
+import { BaseChartDirective, Color, Label } from 'ng2-charts';
+import { finalize } from 'rxjs/operators';
+import { DatePipe } from '@angular/common';
+import { WeatherResponse } from '../../model/weather/weather-response';
+import { ResponseStatus } from '../../model/response-status.enum';
 
 @Component({
   selector: 'app-weather-hourly-chart',
   templateUrl: './weather-hourly-chart.component.html',
-  styleUrls: ['./weather-hourly-chart.component.css']
+  styleUrls: ['./weather-hourly-chart.component.css'],
 })
 export class WeatherHourlyChartComponent implements OnInit {
-
   @Input() hourlyWeatherResponse: HourlyWeatherResponse;
   @Input() city: string;
   @Input() coord: Coord;
   @Input() dataLength = 24;
-  @Output() notifyHourlyWeatherResponse: EventEmitter<HourlyWeatherResponse> = new EventEmitter<HourlyWeatherResponse>();
+  @Output() notifyHourlyWeatherResponse: EventEmitter<HourlyWeatherResponse> =
+    new EventEmitter<HourlyWeatherResponse>();
 
   hourlyWeatherResponseLoading = true;
   errorMessage: string;
 
-  constructor(
-    private weatherService: WeatherService
-  ) { }
+  constructor(private weatherService: WeatherService) {}
 
   public lineChartData: ChartDataSets[] = [];
   public lineChartLabels: Label[];
-  public lineChartOptions: (ChartOptions) = {
+  public lineChartOptions: ChartOptions = {
     maintainAspectRatio: false,
     responsive: true,
     legend: {
-      labels: { fontColor: 'white' }
+      labels: { fontColor: 'white' },
     },
     scales: {
-      xAxes: [{
-        ticks: {
-          fontColor: 'white',
-          fontSize: 11,
-          stepSize: 1,
+      xAxes: [
+        {
+          ticks: {
+            fontColor: 'white',
+            fontSize: 11,
+            stepSize: 1,
+          },
+          gridLines: { color: 'rgba(255,255,255,0.1)' },
         },
-        gridLines: { color: 'rgba(255,255,255,0.1)' }
-      }],
+      ],
       yAxes: [
         {
           id: 'y-axis-0',
@@ -57,11 +57,9 @@ export class WeatherHourlyChartComponent implements OnInit {
             beginAtZero: true,
             fontColor: 'white',
             fontSize: 15,
-            callback: (value, index, values) => {
-              return value + '°';
-            }
+            callback: (value, index, values) => value + '°',
           },
-          gridLines: { color: 'rgba(255,255,255,0.1)' }
+          gridLines: { color: 'rgba(255,255,255,0.1)' },
         },
         {
           id: 'y-axis-1',
@@ -72,37 +70,37 @@ export class WeatherHourlyChartComponent implements OnInit {
             beginAtZero: true,
             fontColor: 'white',
             fontSize: 10,
-            callback: (value, index, values) => {
-              return ((value as number) * 100) + '%';
-            }
+            callback: (value, index, values) => (value as number) * 100 + '%',
           },
         },
-      ]
+      ],
     },
     layout: {
       padding: {
         left: 0,
         right: 0,
         top: 0,
-        bottom: 15
-      }
-    }
+        bottom: 15,
+      },
+    },
   };
 
   images: string[];
-  public lineChartPlugins = [{
-    afterDraw: chart => {
-      const ctx = chart.chart.ctx;
-      const xAxis = chart.scales['x-axis-0'];
-      const yAxis = chart.scales['y-axis-0'];
-      xAxis.ticks.forEach((value, index) => {
-        const x = xAxis.getPixelForTick(index);
-        const image = new Image();
-        image.src = this.images[index];
-        ctx.drawImage(image, x - 9, yAxis.bottom + 25, 20, 20);
-      });
-    }
-  }];
+  public lineChartPlugins = [
+    {
+      afterDraw: (chart) => {
+        const ctx = chart.chart.ctx;
+        const xAxis = chart.scales['x-axis-0'];
+        const yAxis = chart.scales['y-axis-0'];
+        xAxis.ticks.forEach((value, index) => {
+          const x = xAxis.getPixelForTick(index);
+          const image = new Image();
+          image.src = this.images[index];
+          ctx.drawImage(image, x - 9, yAxis.bottom + 25, 20, 20);
+        });
+      },
+    },
+  ];
 
   public lineChartColors: Color[] = [
     {
@@ -111,7 +109,7 @@ export class WeatherHourlyChartComponent implements OnInit {
       pointBackgroundColor: 'rgb(63,165,255)',
       pointBorderColor: '#fff',
       pointHoverBackgroundColor: '#fff',
-      pointHoverBorderColor: 'rgb(63,165,255)'
+      pointHoverBorderColor: 'rgb(63,165,255)',
     },
     {
       backgroundColor: 'rgba(205,229,255,0.6)',
@@ -119,7 +117,7 @@ export class WeatherHourlyChartComponent implements OnInit {
       pointBackgroundColor: 'rgba(205,229,255)',
       pointBorderColor: '#fff',
       pointHoverBackgroundColor: '#fff',
-      pointHoverBorderColor: 'rgba(205,229,255)'
+      pointHoverBorderColor: 'rgba(205,229,255)',
     },
   ];
   public lineChartLegend = true;
@@ -139,11 +137,10 @@ export class WeatherHourlyChartComponent implements OnInit {
     }
   }
 
-
   updateHourlyWeather(coord: Coord): void {
     const hourlyWeatherResponse$ = this.weatherService.getHourlyWeatherResponseByCoord(coord);
-    hourlyWeatherResponse$
-      .subscribe(r => {
+    hourlyWeatherResponse$.subscribe(
+      (r) => {
         if (r.responseStatus.toString() !== 'SUCCESS') {
           this.hourlyWeatherResponseLoading = true;
           this.errorMessage = r.responseMessage;
@@ -153,33 +150,29 @@ export class WeatherHourlyChartComponent implements OnInit {
         this.notifyHourlyWeatherResponse.emit(r);
         this.prepareChart(r);
         this.hourlyWeatherResponseLoading = false;
-      }, error => {
+      },
+      (error) => {
         this.hourlyWeatherResponseLoading = true;
         this.errorMessage = error.message;
         setTimeout(this.updateHourlyWeather.bind(this), 6000);
-      });
+      }
+    );
   }
 
   prepareChart(hourlyWeatherResponse: HourlyWeatherResponse): void {
     const weathers = hourlyWeatherResponse.hourlyWeather.weathers.slice(0, this.dataLength);
     const temp: ChartDataSets = { data: [], label: 'Temperature' };
     const pop: ChartDataSets = { data: [], label: 'Probability of precipitation', type: 'bar', yAxisID: 'y-axis-1' };
-    temp.data = weathers.map(weather => {
-      return parseFloat((weather.temp - 273.15).toFixed(1));
-    });
-    pop.data = weathers.map(weather => {
-      return weather.pop;
-    });
+    temp.data = weathers.map((weather) => parseFloat((weather.temp - 273.15).toFixed(1)));
+    pop.data = weathers.map((weather) => weather.pop);
     this.lineChartLabels = weathers.map((weather, index) => {
       if (index === 0) {
         return 'now';
       }
-      return new DatePipe('en-US').transform(weather.dt * 1000 , 'H');
+      return new DatePipe('en-US').transform(weather.dt * 1000, 'H');
     });
 
-    this.images = weathers.map(weather => {
-      return 'https://openweathermap.org/img/wn/' + weather.icon + '.png';
-    });
+    this.images = weathers.map((weather) => 'https://openweathermap.org/img/wn/' + weather.icon + '.png');
     this.lineChartData.push(temp, pop);
   }
 }
