@@ -1,16 +1,13 @@
-import {Observable, of} from 'rxjs';
-import {WeatherResponse} from '../model/weather/weather-response';
-import {DailyWeatherResponse} from '../model/weather/daily-weather-response';
-import {HttpClient} from '@angular/common/http';
-import {Coord} from '../model/weather/coord';
-import {JsonObject} from '@angular/compiler-cli/ngcc/src/packages/entry_point';
-import {Injectable} from '@angular/core';
-import {HourlyWeatherResponse} from '../model/weather/hourly-weather-response';
-import {catchError} from 'rxjs/operators';
-import {environment} from '../../environments/environment';
+import { Observable } from 'rxjs';
+import { WeatherResponse } from '../model/weather/weather-response';
+import { DailyWeatherResponse } from '../model/weather/daily-weather-response';
+import { HttpClient } from '@angular/common/http';
+import { Coord } from '../model/weather/coord';
+import { Injectable } from '@angular/core';
+import { HourlyWeatherResponse } from '../model/weather/hourly-weather-response';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class WeatherService {
   private baseUrl = 'http://localhost/api/';
@@ -46,33 +43,26 @@ export class WeatherService {
     //     });
     // });
 
-    return this.http
-      .get<WeatherResponse>(
-        this.baseUrl + 'weather/city/' + city + '/' + this.piServiceApiKey);
+    return this.http.get<WeatherResponse>(this.baseUrl + 'weather/city/' + city + '/' + this.piServiceApiKey);
   }
 
   public getDailyWeatherResponseByCoord(coord: Coord): Observable<DailyWeatherResponse> {
-    return this.http
-      .get<DailyWeatherResponse>(
-        this.baseUrl + 'dailyWeather/coord/' + coord.lat + '/' + coord.lon + '/' + this.piServiceApiKey);
+    return this.http.get<DailyWeatherResponse>(
+      this.baseUrl + 'dailyWeather/coord/' + coord.lat + '/' + coord.lon + '/' + this.piServiceApiKey
+    );
   }
 
   public getHourlyWeatherResponseByCoord(coord: Coord): Observable<HourlyWeatherResponse> {
-    return this.http
-      .get<HourlyWeatherResponse>(
-        this.baseUrl + 'hourlyWeather/coord/' + coord.lat + '/' + coord.lon + '/' + this.piServiceApiKey);
+    return this.http.get<HourlyWeatherResponse>(
+      this.baseUrl + 'hourlyWeather/coord/' + coord.lat + '/' + coord.lon + '/' + this.piServiceApiKey
+    );
   }
 
   public getLocationByIP(ipAddr = ''): Observable<string> {
     if (ipAddr === '') {
-      return this.http
-        .get<string>(this.baseUrl + 'ipGeolocation/' + this.piServiceApiKey);
+      return this.http.get<string>(this.baseUrl + 'ipGeolocation/' + this.piServiceApiKey);
     } else {
-      return this.http
-        .get<string>(this.baseUrl + 'ipGeolocation/' + ipAddr + '/' + this.piServiceApiKey);
+      return this.http.get<string>(this.baseUrl + 'ipGeolocation/' + ipAddr + '/' + this.piServiceApiKey);
     }
-
   }
-
-
 }
