@@ -1,15 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import {Coord} from '../model/weather/coord';
-import {WeatherService} from './weather.service';
-import {WeatherResponse} from '../model/weather/weather-response';
-import {DailyWeatherResponse} from '../model/weather/daily-weather-response';
-import {HourlyWeatherResponse} from '../model/weather/hourly-weather-response';
+import { Coord } from '../model/weather/coord';
+import { WeatherService } from './weather.service';
+import { WeatherResponse } from '../model/weather/weather-response';
+import { DailyWeatherResponse } from '../model/weather/daily-weather-response';
+import { HourlyWeatherResponse } from '../model/weather/hourly-weather-response';
 
 @Component({
   selector: 'app-weather',
   templateUrl: './weather.component.html',
-  styleUrls: ['./weather.component.css']
+  styleUrls: ['./weather.component.css'],
 })
 export class WeatherComponent implements OnInit {
   city = undefined;
@@ -26,15 +26,11 @@ export class WeatherComponent implements OnInit {
   currentPage = 0;
   currentTimer = 0;
 
-  constructor(
-    private weatherService: WeatherService,
-    private route: ActivatedRoute
-  ) { }
+  constructor(private weatherService: WeatherService, private route: ActivatedRoute) {}
 
   ngOnInit(): void {
-
     // It enables customized width. If a specify display cannot show full width, this may solve the problem.
-    this.route.queryParams.subscribe(params => {
+    this.route.queryParams.subscribe((params) => {
       if (!isNaN(parseInt(params.w, 10))) {
         document.body.style.width = params.w + 'px';
       }
@@ -51,13 +47,12 @@ export class WeatherComponent implements OnInit {
         }
       }
     }, 10);
-
   }
 
   updateCityName(): void {
     const jsonResponse$ = this.weatherService.getLocationByIP();
     jsonResponse$.subscribe(
-      response => {
+      (response) => {
         const data = JSON.stringify(response);
         const jsonResponse = JSON.parse(data);
         if (jsonResponse === null || jsonResponse === undefined || jsonResponse.ipGeolocation === undefined) {
@@ -79,11 +74,12 @@ export class WeatherComponent implements OnInit {
         }, 1000 * 60 * 30);
         this.locationLoading = false;
         this.errorMessage = undefined;
-    },
-        error => {
+      },
+      (error) => {
         this.errorMessage = '[Error Message]: ' + error.message;
         setTimeout(this.updateCityName.bind(this), 10000);
-    });
+      }
+    );
   }
 
   setCoord(coord: Coord): void {
